Separate azure login bits in LocalServer + Helper (mocked part)
